Add lost bed contract test
